% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cep2coo.R
\name{cep2coo}
\alias{cep2coo}
\title{From CEP to geographic coordinate system}
\usage{
cep2coo(cep)
}
\arguments{
\item{cep}{Number with 8 digits}
}
\description{
Transforme CEP in geographic coordinate system
}
\examples{
NULL

}
