% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Crosstabs}
\usage{
crosstab(data, varX, varY, categ)
}
\arguments{
\item{data}{Dataset}

\item{varX}{Variable}

\item{varY}{Variable}

\item{categ}{Choose category from y variable to analyse}
}
\description{
Perform crosstabs
}
\examples{
NULL

}
