% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab2.R
\name{crosstab2}
\alias{crosstab2}
\title{Crosstabs to use with purrr::map}
\usage{
crosstab2(data, varX, varY, categ)
}
\arguments{
\item{data}{Dataset}

\item{varX}{Variable}

\item{varY}{Variable}

\item{categ}{Choose category from y variable to analyse}
}
\description{
Perform crosstabs to use with purrr::map
}
\examples{
NULL

}
