% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq2.R
\name{freq2}
\alias{freq2}
\title{Frequency to use with purrr::map}
\usage{
freq2(data, var)
}
\arguments{
\item{data}{Dataset}

\item{var}{Variable}
}
\description{
Perform frequencies to use with purrr::map
}
\examples{
NULL

}
