% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOutlier.R
\name{isOutlier}
\alias{isOutlier}
\title{Is a outlier}
\usage{
isOutlier(x, a)
}
\arguments{
\item{x}{Variable}

\item{a}{Multiplicative factor (a * interquartile range)}
}
\description{
In a distribution is defined whether a value is a outlier
}
\examples{
NULL

}
