% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Crosstabs}
\usage{
crosstab(data, x, y, simulate.p.value = FALSE)
}
\arguments{
\item{data}{Dataset}

\item{x}{Variable x}

\item{y}{Variable y}

\item{simulate.p.value}{Use p-value simulated}
}
\description{
Perform crosstabs
}
\examples{
NULL

}
